Refactor SwanLabKey to separate module and update tests
Moved the SwanLabKey class from exp.py to a new key.py module, improving code organization and encapsulation. Updated SwanLabExp to use the new SwanLabKey signature. Added get_class method to DataWrapper for type consistency checks. Refactored unit tests to use UseMockRunState context manager for better test isolation and reliability.
